requirements.txt
# ---------------- Core scientific stack ----------------
numpy>=1.23          # Fundamental array operations
torch>=2.1.2         # PyTorch deep‑learning framework (CUDA / CPU)
torchvision>=0.16.2  # Vision utilities + pretrained models
torchaudio>=2.1.2    # (Optional) audio utilities, share Torch version

# --------------- RAW decoding & image I/O --------------
rawpy>=0.18          # LibRaw Python binding for Sony ARW / Nikon NEF
opencv-python>=4.9   # Fast image reading / color conversions
Pillow>=10.0         # Basic image loading / saving (JPEG, TIFF)

# ---------------- Data augmentation --------------------
albumentations>=1.3  # Powerful CV augmentations (crop, noise, WB shift)
imgaug>=0.4          # Extra augmentation ops (geometric / pixel)
colour-science>=0.4.4  # Color science utilities (tone curves, color space)

# ---------------- Utility & logging --------------------
tqdm>=4.66           # Neat progress bars for preprocessing / training
pyyaml>=6.0          # Read YAML configs
typing-extensions>=4.10  # Forward‑compat typing helpers

# --------------- Experiment tracking -------------------
tensorboard>=2.14    # Loss & metric visualisation
dvc[gs]>=3.43        # Data Version Control + Google Storage remote
dvc[s3]>=3.43        # (Optional) S3 remote support

# ---------------- Evaluation / metrics -----------------
scikit-image>=0.22   # PSNR / SSIM implementations
scikit-learn>=1.3    # Additional utilities (e.g., LPIPS helpers)

# ---------------- Testing / CI -------------------------
pytest>=7.4          # Unit & integration tests
pytest-cov>=4.1      # Coverage reports for CI